move xarray read/load functions
