+++ tests/datasets/pvnet/test_materialise.py
@@ -3,7 +3,7 @@ import numpy as np
 import xarray as xr
 
 from ocf_data_sampler.common.tensorstore import open_zarr
-from ocf_data_sampler.utils import load, load_data_dict
+from ocf_data_sampler.datasets.pvnet.materialise import load, load_data_dict
 
 
 def test_load():
